Made several changes
Moved the variables to a JSON file that is easier to read and edit.

Added the option to set all forms to false.

Parameters are checked to make sure that they exist and are correct.

Output file can now adjust itself based on the longest name of all Pokémon that are being written to file.

Changed the names of some files.